cloud_functions/harvest/projects/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.9
    # via data-pipeline-tools
aiosignal==1.3.1
    # via aiohttp
attrs==22.2.0
    # via
    #   aiohttp
    #   cattrs
    #   requests-cache
cachetools==5.3.0
    # via google-auth
cattrs==23.1.2
    # via requests-cache
certifi==2023.7.22
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.7
    # via pyforecast
data-pipeline-tools @ git+https://github.com/tpximpact/data-pipeline-tools@b4fd4630bcc06a7ac06e0b5cbb2917b47c23b860
    # via harvest-projects (pyproject.toml)
db-dtypes==1.0.5
    # via pandas-gbq
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.23.0
    # via
    #   google-api-python-client
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   pandas-gbq
google-api-python-client==2.154.0
    # via data-pipeline-tools
google-auth==2.16.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   pandas-gbq
    #   pydata-google-auth
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.0.0
    # via
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery==3.27.0
    # via
    #   data-pipeline-tools
    #   pandas-gbq
google-cloud-core==2.4.1
    # via
    #   data-pipeline-tools
    #   google-cloud-bigquery
google-cloud-secret-manager==2.21.1
    # via data-pipeline-tools
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.1
    # via google-cloud-bigquery
googleapis-common-protos==1.58.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.6
    # via google-cloud-secret-manager
grpcio==1.52.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.52.0
    # via google-api-core
holidays==0.62
    # via data-pipeline-tools
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via
    #   requests
    #   yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.24.2
    # via
    #   db-dtypes
    #   pandas
    #   pandas-gbq
    #   pyarrow
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.0
    # via
    #   db-dtypes
    #   google-cloud-bigquery
    #   pandas-gbq
pandas==1.5.3
    # via
    #   db-dtypes
    #   pandas-gbq
pandas-gbq==0.24.0
    # via data-pipeline-tools
platformdirs==4.3.6
    # via requests-cache
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.25.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
protobuf==4.22.0
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==11.0.0
    # via
    #   db-dtypes
    #   pandas-gbq
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydata-google-auth==1.7.0
    # via pandas-gbq
pyforecast @ git+https://github.com/tpximpact/py_forecast@c90b2dbc0105e51621c0303dbd6fd9e17a784961
    # via data-pipeline-tools
pyparsing==3.2.0
    # via httplib2
python-dateutil==2.8.2
    # via
    #   google-cloud-bigquery
    #   holidays
    #   pandas
pytz==2022.7.1
    # via pandas
requests==2.28.2
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   pyforecast
    #   requests-cache
    #   requests-oauthlib
requests-cache==1.2.1
    # via pyforecast
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
setuptools==67.6.1
    # via
    #   pandas-gbq
    #   pydata-google-auth
six==1.16.0
    # via
    #   google-auth
    #   python-dateutil
    #   url-normalize
uritemplate==4.1.1
    # via google-api-python-client
url-normalize==1.4.3
    # via requests-cache
urllib3==1.26.14
    # via
    #   requests
    #   requests-cache
yarl==1.18.3
    # via aiohttp
